Base for usage of multiple instance for API versions
