+++ src/test/index.ts
@@ -28,6 +28,11 @@ commander
     console.log(
       await client.get('https://api.twitter.com/1.1/search/tweets.json?q=@alkihis')
     );
+
+    console.log(
+      'With v1',
+      await client.v1.get('search/tweets.json?q=@alkihis')
+    );
   }
   else if (commander.appOnly) {
     // OAuth2
@@ -40,6 +45,12 @@ commander
     console.log(
       'Tweets',
       await client.get('https://api.twitter.com/2/tweets?ids=20,1306166445135605761&expansions=author_id&tweet.fields=public_metrics&user.fields=name,public_metrics')
+    );
+
+    // Tweets
+    console.log(
+      'Tweets with v2',
+      await client.v2.get('tweets?ids=20,1306166445135605761&expansions=author_id&tweet.fields=public_metrics&user.fields=name,public_metrics')
     );
   }
   else if (commander.requestToken) {
